clone/requirements.txt
Django==4.1
djangorestframework==3.14.0
django-crontab==0.7.1
python-decouple==3.8
Requests==2.31.0
